migrate selection types
